switching to slash commands
