+++ package.json
@@ -10,7 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "discord.js": "^13.6.0",
+    "discord.js": "^14.7.1",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "nodemon": "^2.0.20"
